Decompose monolithic conditions.py into modular apts/conditions package.
